requirements.txt
numpy==1.14.3
pandas==0.23.0
psycopg2-binary==2.7.4
SQLAlchemy==1.2.3
tqdm==4.19.6
python-box==3.1.1
requests==2.19.1
tensorflow-gpu==1.8.0
keras==2.1.6